Refactor FacebookContainer to use the new redux setup
